Refactored Entities to use Attributes instead of Annotations, Added User / Weapon directories for Entities
