+++ config/packages/doctrine.yaml
@@ -15,6 +15,7 @@ doctrine:
                 dir: '%kernel.project_dir%/src/Entity'
                 prefix: 'App\Entity'
                 alias: App
+                type: attribute
 
 when@test:
     doctrine:
