+++ config/packages/security.yaml
@@ -9,7 +9,7 @@ security:
     # used to reload user from session & other features (e.g. switch_user)
     app_user_provider:
       entity:
-        class: App\Entity\User
+        class: App\Entity\User\User
         property: username
   role_hierarchy:
     ROLE_ADMIN: [ ROLE_ADMIN ]
