+++ src/Form/WeaponsType.php
@@ -2,7 +2,7 @@
 
 namespace App\Form;
 
-use App\Entity\Weapon;
+use App\Entity\Weapon\Weapon;
 use Doctrine\ORM\EntityRepository;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
